Extract shareable logic to PirDashboardStateProvider.kt
